Refactor request handling and UI components for sent requests
- Updated RequestRow component to display code without prefix.
- Replaced SendRequestsList component with RequestsList for better reusability.
- Integrated API call to fetch sent requests data in SentList component.
- Removed hardcoded request items and replaced with dynamic data from API.
- Created a new RequestsList component to handle rendering of request rows and filters.
- Improved filter handling for family relationships and request statuses.
